requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
rembg>=2.0.50
onnxruntime>=1.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
flask>=3.0.0
python-docx>=0.8.11
gunicorn>=21.2.0
werkzeug>=3.0.0
